backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
pymongo==4.6.0
motor==3.3.2
beanie==1.23.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client for external services
httpx==0.25.2

# Email services
aiosmtplib==3.0.1
jinja2==3.1.2

# SMS services (Twilio)
twilio==8.10.3

# Logging & Monitoring
python-json-logger==2.0.7
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # for testing async clients
faker==20.1.0

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8
